+++ tests/test_qscore.py
@@ -1,50 +1,37 @@
 """Tests for Qscore estimation"""
 
 from iqm.benchmarks.optimization.qscore import *
-
-
-backend = "IQMFakeAdonis"
+from iqm.qiskit_iqm.fake_backends.fake_apollo import IQMFakeApollo
+from iqm.qiskit_iqm.fake_backends.fake_deneb import IQMFakeDeneb
 
 
 class TestQScore:
-    def test_qscore_crystal(self):
+    backend = IQMFakeApollo()
+    qpu_topology = "crystal"
+    custom_qubits_array = [[0, 1, 3], [0, 1, 2, 3] , [0, 1, 2, 3, 4]]
+
+    def test_qscore(self):
         EXAMPLE_QSCORE = QScoreConfiguration(
             num_instances=2,
             num_qaoa_layers=1,
             shots=4,
             calset_id=None,  # calibration set ID, default is None
             min_num_nodes=2,
-            max_num_nodes=None,
+            max_num_nodes=4,
             use_virtual_node=True,
             use_classically_optimized_angles=True,
             choose_qubits_routine="custom",
-            custom_qubits_array=[[2], [2, 0], [2, 0, 1], [2, 0, 1, 3], [2, 0, 1, 3, 4]],
+            custom_qubits_array=self.custom_qubits_array,
             seed=1,
             REM=True,
             mit_shots=10,
-            qpu_topology="crystal"
+            qpu_topology=self.qpu_topology
         )
-        benchmark = QScoreBenchmark(backend, EXAMPLE_QSCORE)
+        benchmark = QScoreBenchmark(self.backend, EXAMPLE_QSCORE)
         benchmark.run()
         benchmark.analyze()
 
-    def test_qscore_star(self):
-        EXAMPLE_QSCORE = QScoreConfiguration(
-            num_instances=2,
-            num_qaoa_layers=1,
-            shots=4,
-            calset_id=None,  # calibration set ID, default is None
-            min_num_nodes=2,
-            max_num_nodes=None,
-            use_virtual_node=True,
-            use_classically_optimized_angles=True,
-            choose_qubits_routine="custom",
-            custom_qubits_array=[[1], [1, 2], [1, 2, 3], [1, 2, 3, 4], [1, 2, 3, 4, 5], [1, 2, 3, 4, 5, 6]],
-            seed=1,
-            REM=True,
-            mit_shots=10,
-            qpu_topology="star"
-        )
-        benchmark = QScoreBenchmark(backend, EXAMPLE_QSCORE)
-        benchmark.run()
-        benchmark.analyze()
+class TestQScoreDeneb(TestQScore):
+    backend = IQMFakeDeneb()
+    qpu_topology = "star"
+    custom_qubits_array = [[1], [1, 2], [1, 2, 3, 4], [1, 2, 3, 4, 5]]